+++ src/components/providers/forms/hooks/index.ts
@@ -17,4 +17,5 @@ export { useOmoModelSource } from "./useOmoModelSource";
 export { useOpencodeFormState } from "./useOpencodeFormState";
 export { useOmoDraftState } from "./useOmoDraftState";
 export { useOpenclawFormState } from "./useOpenclawFormState";
+export { useHermesFormState } from "./useHermesFormState";
 export { useXaiOauth } from "./useXaiOauth";